Add image recipes to support initramfs.
By default build process uses "obmc-phosphor-initramfs" because it requires "conf/machine/include/obmc-bsp-common.inc" in local.conf.
Please add INITRAMFS_IMAGE = "aspeed-image-initramfs" in local.conf to make build process use aspeed-image-initramfs.
It is a work around solution, we will find the solution to fix it.

The ast2600-default machine config file set the KERNEL_DEVICETREE to "aspeed-ast2600-obmc.dtb" for OpenBMC.
We need to find the solution to change it to "aspeed-ast2600-evb.dtb" if user only build the aspeed-image-norootfs.

Change-Id: I91e28f1993486f3d73c5f4c547f75951bf9d748d