Fix secure boot image build failed
[Root Cause]
Exception at the latest version of socsec tool
socsec/__init__.py", line 61$
 in _rsa_bit_length
    key_bit_length = bitarray(bin(rsa_key.n)[2:]).length()
AttributeError: 'bitarray.bitarray' object has no attribute 'length'
[Solution]
So far the version of python3-bitarray is 2.1.0, change to 1.9.2 to fix this issue.
It is a workaround solution.
Once this issue is fixed in socsec tool, we should chage to use the new version of python3-bitarray.

Signed-off-by: Jamin Lin <jamin_lin@aspeedtech.com>
Change-Id: I277c06aad1f2e5df8900d67a090be8bfab606eaf