os-release: Set BUILD_ID to DATETIME
In the community it was discussed having a field that would contain the
date of when an image was built (timestamp) so that it could be used to
generate a unique version id during firmware updates, and that could
also be used to display the specific date of a build such as when it was
released or the general availability date.

The yocto recipe for os-release sets BUILD_ID to DATETIME by default.
The current implementation of BUILD_ID duplicates the output of VERSION,
therefore remove the openbmc implementation of BUILD_ID to pick up the
default value of DATETIME.

The value can be overwritten in a machine's conf file, or set to what it
was before in an os-release.bbappend in a meta layer.

Tested:
- The os-release BUILD_ID has a timestamp value by default, ex:
BUILD_ID="20211025151654".
- The os-release BUILD_ID is set to the value specified in a machine conf
  file, ex:
Ex: In: meta-ibm/conf/machine/p10bmc.conf
BUILD_ID = "20211231120000"
- The os-release BUILD_ID is set to the value specified in a meta-layer
  os-release bbappend:
Ex: In: meta-ibm/recipes-core/os-release/os-release.bbappend
BUILD_ID = "${@run_git(d, 'describe --abbrev=0')}"

Change-Id: If5238c53e0c3463c181788afc713c560fb5fe034
Signed-off-by: Adriana Kobylak <anoo@us.ibm.com>