meta-xilinx: remove subtree
The meta-xilinx layer was used for a now-deleted EVB.  Neither the EVB
nor the meta-xilinx layer have been updated for the Yocto override
syntax change and the meta-xilinx still doesn't have a hardknott or
honister branch (or corresponding support).

I've asked the Xilinx maintainer back in May on when a hardknott
version would be supported and I was told "about a month from now".  I
followed up in August and was told "work is in progress".  As of today
there are still zero commits in meta-xilinx since January 2021.

As such, I do not believe this layer is well-maintained and we have no
specific use for it anymore.  Remove it until someone finds a good
reason to include it and the upstream shows signs of life.

Signed-off-by: Patrick Williams <patrick@stwcx.xyz>
Change-Id: Id14ea55db2ac2779edf42e63cb57ad7d25172ad5