meta-phosphor: remove obmc-system-mgmt from qemu
It was reported that after 5a5f33c729e6b5869362172b63595422eb84a418 the
qemu images are not buildable.  Treat qemu systems like an 'evb' so that
the obmc-system-mgmt package is not included.  Make a minor change to
how evb is specified so that it is treated as a MACHINEOVERRIDE that can
be leveraged in multiple recipes.

Tested by ensuring that `evb-ast2600` still successfully resolves all
package dependencies (with `bitbake -p`).

Signed-off-by: Patrick Williams <patrick@stwcx.xyz>
Change-Id: Iff4573aa3d4aac30a6681ed75741a6e351bda982