obmc-console: Re-introduce OBMC_CONSOLE_HOST_TTY variable
This change re-adds the OBMC_CONSOLE_HOST_TTY configuration variable to
define the host TTY that the obmc-consoler-server should attach to. This
allows individual machines to use a TTY different from the default
ttyVUART0.

Follow-up changes can then remove duplicate code in machine meta layers
that work around the missing support for OBMC_CONSOLE_HOST_TTY.

Tested: Ran a build for both cases with obmc-console.conf and a
server.ttyS2.conf. Checked that the files got installed in the correct
place with the expected names in both cases.

Signed-off-by: Oskar Senft <osk@google.com>
Change-Id: I4fc408577827a827ee10de4f5757726a1eec4a66