phosphor: debug-collector: Limit collection to oops
Currently we trigger the phosphor-debug-collector application to create
a dump whenever a file is present in the pstore directory. This means we
create a dump if eg. a console or ftrace log is captured by ramoops,
which is not indicative of an error.

For example, this is the content of the ramoops filesystem after a
normal reboot followed by a crash:

 console-ramoops-0  dmesg-ramoops-0    dmesg-ramoops-1

Change to ConditionPathExistsGlob and look for anything with the pattern
dmesg-ramoops-*.

This is a better fix for the workaround implemented in 17e10e646324
("meta-ibm:pstore: remove console from ramoops"), allowing systems to
collect more traces without invoking a dump.

Change-Id: If1d93ee86a2cfb7f0902989cb78d128e82578402
Signed-off-by: Joel Stanley <joel@jms.id.au>