meta-ethanolx: Correct 'obmc-console' configuration file
In the AMD EthanolX board CPU0 UART1 is connected to the BMC UART0.
Therefore from the point of BMC UART0 is an upstream (host) UART.
'obmc-console' application gets this information from its configuration
file name ('server.ttyS0.conf').
On the other case setting 'local-tty' inside the configuration file is
intended for the downstream UART settings, which is not present on the
AMD EthanolX board. Therefore remove this setting.

Signed-off-by: Konstantin Aladyshev <aladyshev22@gmail.com>
Change-Id: I3daa1528cb5a7b5ff748189f29da3c3d91899ca4