meta-facebook: add fb-phosphor-image.inc
Builds done using facebook/openbmc do not use the 'obmc-phosphor-image'
image name, but instead use 'MACHINE-image'.  This means that bbappend
files have no effect and we need a different method to get consistent
sets of include files.  Introduce 'fb-phosphor-image.inc' and
'fb-MACHINE-phosphor-image.inc' as the replacements for the
obmc-phosphor-image.bbappend.

Signed-off-by: Patrick Williams <patrick@stwcx.xyz>
Change-Id: I166342c1744d1cd0237f8c75d90d87153fb19b35