run-repotest: do not allow dts files in meta layers
A patch came through recently that had a kernel dts file being patched
in via a meta layer bbappend. Lets have repotest automatically flag
these so we don't have to in reviews.

The OpenBMC project carries a fork of the Linux kernel in openbmc/linux
and that's where all dts files should go. This ensure the upstream
process is moving forward with the changes.

Tested:
- Added a dts file and verified the script failed due to it

Signed-off-by: Andrew Geissler <geissonator@yahoo.com>
Change-Id: Ie82b4d104b73cc18e4c1e0fba772eb92c96ed3fa