meta-quanta: gsj: Remove mac-address daemon
mac-address bitbake recipe in meta-quanta/meta-common points to a
quanta-bmc github. This is not allowed on our meta layer projects,
especially on our "first class machine" like gsj which is part of our CI
process. Remove from gsj for now, consider removing from meta-common
completely if it's not moved out of quanta-bmc.

Signed-off-by: Brandon Kim <brandonkim@google.com>
Change-Id: Ic2d97450abe8d3f532e6ad9b66a98430ce6bf3a5