meta-quanta: s6q: enable sync-mac and nic-ethtool
This commit enables sync-mac and nic-ethtool in the phosphor-network.
The "sync-mac" function can initialize the mac address on first boot, if it can find the object in the inventory.
When phosphor-network triggers systemd to set the mac address, it must be set together with the interface enslaved by bond.
And "nic-ethtool" provides the necessary data through network interface analysis.

Tested:
Check the network object if it worked.
~# busctl introspect xyz.openbmc_project.Network /xyz/openbmc_project/network/bond0

Signed-off-by: Jerry Wan <jerry.wan@quantatw.com>
Change-Id: I63d80b4491f3aa42cae401658a189acaaddb592f