Add back md4 to the build
openldap, on purpose or not, now required md4 support to compile.  While
this is not ideal that we have this dependency, it's easy enough to
re-enable md4 when ldap is enabled.

Unfortunately, the intent of the obmc-user-mgmt-ldap feature looks like
it was broken at some point, and even with that option disabled,
openldap is still included in image builds, so at the moment, md4 is
universally required for all builds.  None of this is ideal, and in the
future, ideally we should:
1. upstream a patch to openldap to not require MD4 support
2. fir the obmc-user-mgmt-ldap feature to meet its intent, and allow
   ldap to be optional in the build.

Tested:  Code builds further with the subtree updates.

Signed-off-by: Ed Tanous <edtanous@google.com>
Change-Id: I3ebf231b21af01df98ca4c11eb496427ea3f9443