meta-ampere: use watchdog status to check A/C power
With the removal of u-boot patch that adds resetreason to /proc/cmdline
no information to check if the BMC is booted from A/C power or cold reboot.
This commit removes the dependency of resetreason on the
phosphor-discover-system-state application and updates the A/C power
check in the ampere_platform_init.sh script to use information from
/sys/class/watchdog/watchdog0/bootstatus.

Tested:
1. Check power restore policy feature work well.
2. Check ampere_platform_init.sh works properly in both A/C and D/C
power.

Signed-off-by: Thang Q. Nguyen <thang@os.amperecomputing.com>
Change-Id: I4f2e7a241f9ddebf21b2b8c021b9ce87a1c8b26c