meta-ibm: set critical services for IBM systems
This takes the default critical services list from meta-phosphor and
appends on extra services IBM is interested in.

Ideally, there would be an easy way to just append an extra file to the
command line but this is going to need some more thought. For now, just
override the entire file.

Tested:
- Booted witherspoon up in QEMU and verified new json file was loaded
  and that stopping one of the new critical services resulted in a
  dump, error log, and BMC Quiesced state.

Signed-off-by: Andrew Geissler <geissonator@yahoo.com>
Change-Id: Icf5db5331072cbcbd0359e0b789b664e31423124