rsyslog: fix default server config
Directory /etc/rsyslog.d/ is expected to hold several *.conf files each
serving an independent purpose, all processed in order for each syslog
message.

The deprecated action ~ means "discard action is carried out, the
received message is immediately discarded. No further processing of it
occurs". Thus the config files that happen to be alphabetically after
server.conf are not going to ever see any messages, and that is
unexpected and confusing.

Fix by shipping a config which writes any message to /dev/null. This is
needed to avoid error 2103 in default configuration where no other
output actions are specified for rsyslog. Also enable processing of
other *.conf file which might be shipped by other packages or added by
an end user.

This should also be accomplished by a similar change to phosphor-logging
so that it wouldn't revert to wrong behaviour when sending to remote
syslog is disabled.

Signed-off-by: Paul Fertser <fercerpav@gmail.com>
Change-Id: I5da9a46df1b8cf1326bb1a7b93af18cfc568dee8