image_types_phosphor: Add compatible name to MANIFEST
This adds CompatibleNames field to the manifest file which can contain
compatibility strings as described in the Software interface.

The field is set by the OBMC_COMPATIBLE_NAMES variable in the
machine.conf file, e.g.

OBMC_COMPATIBLE_NAMES = "\
    foo.bar \
    baz.bim \
"

Also omit ExtendedVersion from the manifest file if it is not set.

Signed-off-by: Justin Ledford <justinledford@google.com>
Change-Id: Id27b4d6ec25ec7c3d3d5942368f40e618535eafe