phosphor-state-manager: move host symlinks to correct subpkg
The `multi-user.target.requires/...Host@{i}.service` symlinks were
being added to the `obmc-targets` sub-package, but nothing explicitly
depends on the `host` sub-package.  On systems that do not include
the `host` sub-package, this could result in broken symlinks, and
on systems which do not use the targets it can result in the `host`
daemon not starting.

Move these `Host@{i}.service` symlinks into the `host`
sub-package since that is where the original service file resides.

Signed-off-by: Patrick Williams <patrick@stwcx.xyz>
Change-Id: I43aa5fdbe73f0af3c5d213e45e3a02f2dc093d79