meta-phosphor: Fix bios-bmc-smm-error-logger DEPENDS
Add sdbusplus and phosphor-dbus-interfaces to DEPENDS. The new srcrev
bump change includes a newly added meson depedency.

fmt was missed initially but it seemed to be building fine without it.
Just adding it to be safe.

Tested: Built an internal platform that builds this daemon, and saw that
it builds properly.

Signed-off-by: Brandon Kim <brandonkim@google.com>
Change-Id: I160c2eeef07a301c8ace43d410948b2d0d35893c