phosphor-state-manager: srcrev bump 1e91dc6357..b4cbfac10b
Corey-Hardesty (1):
      Host reboot configuration support

Corey-Hardesty (1):
      Define new 'obmc-chassis-blackout@.target'

"obmc-chassis-blackout@.target" was added to ensure the system can
recognize when it has entered a blackout state. This target is to be
called from chassis-state-manager when determining the initial state
of the chassis and finds that it experienced a power blackout.
Chassis-state-manager does this by checking the status of the pgood
pin along with the last known power state of the system to determine
if the target should be applied. If the last known power state was
"powered on" and the pgood pin is false, the chassis experienced a
blackout. To accommodate for any discrepencies across the systemd
services and targets may be added such as resetting host-state-sensors
or powering off leds to ensure the system stays in a predictable state.

Change-Id: I0832cd6300e9a1742c9ec2a9c5f68890b4e11918
Signed-off-by: Andrew Geissler <openbmcbump-github@yahoo.com>
Signed-off-by: Corey Hardesty <corey.hardesty@icloud.com>