intel-pfr-signing-utility: support openssl 3.0
The functions currently used in the sslhelper.c are deprecated in openssl 3.0,
but the APIs necessary for conversion also do not exist in 1.1.
Add a #define on the OPENSSL_VERSION_NUMBER to identify 3.x (or greater)
support and switch between the two API sets.

Apply 0001-support-openssl-3.0.patch to fix build failed with openssl 3.0.

Signed-off-by: Jamin Lin <jamin_lin@aspeedtech.com>