meta-ampere: support ampere power/host already on targets
There are no targets to identity the already on of power state and the
already running of host state. This commit adds those targets.
The targets will be the check points to start ampere services such as
drivers binder, hang detection...
obmc-power-already-on@.target will be always activated when
chassis-power-control is asserted. It will be deactivated before
deasserting chassis-power-control.
obmc-host-already-on@.target will be activated when the host%i-ready
is triggered. The target will be deactivated before turning
off/reseting the host.

Tested:
1. The targets should be activated/deactivated as designing.

Signed-off-by: Thu Nguyen <thu@os.amperecomputing.com>
Change-Id: I09a8b94d8043f31f7cee83b29765f7e2c7a5c35a