meta-google: remove glome-config from glome's RDEPENDS
glome-config only provides the config file for glome. It is more
appropriate to decouple them by adding glome-config to
OBMC_IMAGE_EXTRA_INSTALL in the meta-google layer. Then when glome is
migrated to another meta layer, it won't need this RDEPENDS.
(This is a follow-up of
https://gerrit.openbmc.org/c/openbmc/openbmc/+/56618/)

Also add glome-login to OBMC_IMAGE_EXTRA_INSTALL since it is also
needed.

Tested:
Built an image and run it on a real machine. Verified that glome is
enabled and the generated link gave a valid password.

Signed-off-by: Leo Tu <leotu@google.com>
Change-Id: I985670454f4749c5297261ec81466fed9cdc5c40