treewide: more obmc-host-ctl cleanups
obmc-host-ctl is defined as a MACHINE, DISTRO, and IMAGE feature.
The IMAGE feature pulls in a virtual provider only if the COMBINED
feature (MACHINE + DISTRO) is enabled.  This is only done on openpower
machines.  There are a lot of other places where machines are half-way
disabling this feature and/or changing variables which have zero effect
anyhow.  Clean up almost all references to `obmc-host-ctl`.

Signed-off-by: Patrick Williams <patrick@stwcx.xyz>
Change-Id: I090d782bb6a34dc4e3c43df97e8ccf6e1812fed5