meta-ibm: p10bmc: Add rainier runqemu options
Qemu boots kernel+dtb directly, and mounts wic image (full filesystem)
as an emulated mmc device.

We add zImage as a KERNEL_IMAGETYPES (plural is important) so it is
copied to the deploy directory.

Currently there is only a wic.xz which cannot be booted directly. We add
wic.qcow2 so the wic image is exported as a usable image for Qemu. This
is a sparse image so empty space does not take up disk.

TODO:
  1. qemu requires the image size to be a power of two. Current
     workaround:

     qemu-img resize p10bmc/tmp/deploy/images/p10bmc/*.wic.qcow2 16G

  2. Boot via u-boot. Requires qemu mmc patches from Cédric's tree, and
     creating an image with the boot0/boot1 partitions in it.

Change-Id: Ic64023b2d7f9bce80a025a1d847228e90e0b72bd
Signed-off-by: Joel Stanley <joel@jms.id.au>