libmctp-intel-test: Change to use “0x7c” message type for echo command test
In the previous design, we set the message type to “0x85” to support echo command for large message stress test.
However, according to the MCTP spec, the bit 7 is used for integrity check bit. Besides, it conflict with message type “0x05” for SPDM.
So, changes to use message type "0x7c" for echo command support, please refer to MCTP IDs and Code spec for each message type defined.

Signed-off-by: Jamin Lin <jamin_lin@aspeedtech.com>
Change-Id: I7b5614692de713b5806d66f902f4cdde626ac87b