meta-delta/meta-ahe50dc: Add new machine layer
This is an Open19 power shelf with two AST1250 BMCs in a redundant
active/active configuration.  It's not a server and hence has no host to
manage, and so omits many standard BMC services such as host power
control, serial console, and KVM support.  Its primary purpose is to
manage and monitor the six PSUs and 50 efuses (12VDC power outputs) the
AHE-50DC uses to provide power to Open19 bricks and switches.

Signed-off-by: Zev Weiss <zev@bewilderbeest.net>
Change-Id: I56cd21585b987de8d97a15612457e40b4a214d7d