meta-phosphor: introduce fit-image.bbclass
Extract relevant functions from upstream's kernel-fitimage.bbclass
in order to be able to create a FIT image containing the rootfs
as the initramfs.  This will let us generate MTD images with a single
rootfs-as-initramfs.

My intention is to ensure this is working in meta-phosphor and then
work at moving the bbclass upstream.

Signed-off-by: Patrick Williams <patrick@stwcx.xyz>
Change-Id: I9b996f32be718a69da863aa2f14a8fe8278c84f9