meta-phosphor: phosphor-dbus-interfaces: add explicit task-dep on unpack
The recipe has an extension to the `do_write_config` task that
forces all of the YAML subdirectory options to `false` using the list
in the `meson_options.txt`.  Since we are parsing the
`meson_options.txt`, we need it to be present, so we need to explicitly
rely on the `do_unpack` task.

Tested:

The `log.task_order` has the correct order.
```
$ cat log.task_order
20230113-151838.797418 do_fetch (3200585): log.do_fetch.3200585
20230113-151840.771096 do_unpack (3201820): log.do_unpack.3201820
20230113-151842.762491 do_write_config (3202881): log.do_write_config.3202881
```
Ran `bitbake phosphor-dbus-interfaces -ne` and examined the `taskdeps`
structure to confirm the dependency is added.

Signed-off-by: Patrick Williams <patrick@stwcx.xyz>
Change-Id: I0d81d127fd1bda39eeb38e6ce89614e1f4ddfa30