romed8hm3: Adjust kernel config for nct6779 support
On the romed8hm3 board (as on some other ASRock AMD platforms) the host
CPU's TSI interface is connected to a super-io chip instead of being run
directly to the BMC, so in order to retrieve CPU temperature readings we
have to go through the super-io chip's back-door i2c interface.  This
enables the driver support we need to do so.

Signed-off-by: Zev Weiss <zev@bewilderbeest.net>
Change-Id: I5db6fb9e1a290f42af9561c2eb889ac00cddd104